funcionando

+++ complexivo/admin.php
@@ -26,47 +26,49 @@ $resultado = $sql->fetchAll(PDO::FETCH_ASSOC);
 
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
     <link rel="stylesheet" href="css/styles.css">
+    <link rel="stylesheet" href="css/fontawesome-free-6.3.0-web/css/all.min.css">
+
 </head>
 
 <body>
-<header>
-    <div class="navbar navbar-expand-lg navbar-dark bg-dark ">
-        <div class="container">
-            <a href="#" class="navbar-brand ">
-                <strong>Tienda Online Sublimados Quito</strong>
-            </a>
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarHeader" aria-controls="navbarHeader" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
+    <header>
+        <div class="navbar navbar-expand-lg navbar-dark bg-dark ">
+            <div class="container">
+                <a href="#" class="navbar-brand ">
+                    <strong>Tienda Online Sublimados Quito</strong>
+                </a>
+                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarHeader" aria-controls="navbarHeader" aria-expanded="false" aria-label="Toggle navigation">
+                    <span class="navbar-toggler-icon"></span>
+                </button>
 
-            <div class="collapse navbar-collapse" id="navbarHeader">
-                <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-                    <li class="nav-item">
-                        <a href="admin.php" class="nav-link active">Catalogo</a>
-                    </li>
+                <div class="collapse navbar-collapse" id="navbarHeader">
+                    <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+                        <li class="nav-item">
+                            <a href="admin.php" class="nav-link active">Catalogo</a>
+                        </li>
 
-                    <li class="nav-item">
-                        <a href="contactoAdmin.php" class="nav-link ">Contacto</a>
-                    </li>
-                </ul>
+                        <li class="nav-item">
+                            <a href="contactoAdmin.php" class="nav-link ">Contacto</a>
+                        </li>
+                    </ul>
 
-                <?php if (isset($_SESSION['user_id'])) { ?>
-                    <div class="dropdown">
-                        <button class="btn btn-success btn-sm dropdown-toggle" type="button" id="btn_session" data-bs-toggle="dropdown" aria-expanded="false">
-                        <i class="fa-regular fa-user"></i><?php echo $_SESSION['user_name']; ?></a>
-                        </button>
-                        <ul class="dropdown-menu" aria-labelledby="btn_session">
-                        <li><a class="dropdown-item" href="createProduct.php">Crear Articulo</a></li>
-                            <li><a class="dropdown-item" href="logout.php">Cerrar Sesión</a></li>
-                        </ul>
-                    </div>
-                <?php } else { ?>
-                    <a href="login.php" class="btn btn-success btn-sm "><i class="fa-regular fa-user"></i></i>Ingresar</a>
-                <?php } ?>
+                    <?php if (isset($_SESSION['user_id'])) { ?>
+                        <div class="dropdown">
+                            <button class="btn btn-success btn-sm dropdown-toggle" type="button" id="btn_session" data-bs-toggle="dropdown" aria-expanded="false">
+                                <i class="fa-regular fa-user"></i><?php echo $_SESSION['user_name']; ?></a>
+                            </button>
+                            <ul class="dropdown-menu" aria-labelledby="btn_session">
+                                <li><a class="dropdown-item" href="createProduct.php">Crear Articulo</a></li>
+                                <li><a class="dropdown-item" href="logout.php">Cerrar Sesión</a></li>
+                            </ul>
+                        </div>
+                    <?php } else { ?>
+                        <a href="login.php" class="btn btn-success btn-sm "><i class="fa-regular fa-user"></i></i>Ingresar</a>
+                    <?php } ?>
+                </div>
             </div>
         </div>
-    </div>
-</header>
+    </header>
 
     <!--contenido-->
     <main>
@@ -93,11 +95,9 @@ $resultado = $sql->fetchAll(PDO::FETCH_ASSOC);
                                         <a href="updateProducts.php?id=<?php echo $row['id']; ?>&token=<?php echo hash_hmac('sha1', $row['id'], KEY_TOKEN); ?>" class="btn btn-primary">Editar</a>
                                     </div>
                                     <div class="btn-group">
-                                        <a href="class/eliminarProduct.php?id=<?php echo $row['id']; ?>&token=<?php echo hash_hmac('sha1', $row['id'], KEY_TOKEN); ?>" class="btn btn-danger">Eliminar</a>
+                                    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.1/css/all.min.css">
+                                    <a class="btn btn-danger" href="class/eliminarProduct.php?id=<?php echo $row['id']; ?>&token=<?php echo hash_hmac('sha1', $row['id'], KEY_TOKEN); ?>" class="btn btn-danger" onclick="return confirm('¿Realmente desea eliminar?')"><i class="fas fa-trash mr-2"></i>Eliminar</a>
                                     </div>
-
-
-
                                 </div>
                             </div>
                         </div>
